CSS Message Tab and Badge Counter
I finished coding all of the CSS crap for the Message Inbox Icon, as
well as a CSS Span Badge, which should automatically display the current
amount of unread messages within the inbox live thanks to some AJAX that
I found. And luckily it's already programmed within the CSS to not
display the Badge if the Span is EMPTY.

So now all that's left to do is for someone to write the MVC Razor code
that will input the number between the SPAN tags. But please keep in
mind to make sure that if there are no unread messages, to not put a
zero there - leave it NULL! That's the only way that the "badge:empty"
within the bootstrap CSS will hide it.